refactor: 공용함수 분리
