+++ hooks/README.md
@@ -28,10 +28,10 @@
   - [ ] 유효성 검증 + 상태 변경 함수
   - [ ] 모든 에러가 없는지 변수
 
-- [ ] 유효성 검증 함수
+- [ ] 유효성 검증 함수 분리
 
-  - [ ] 숫자 검증
+  - [x] 숫자 검증
   - [ ] 월 검증
   - [ ] 년 검증
 
-- [ ] 모든 에러가 없는지 체크하는 함수
+- [x] 모든 에러가 없는지 체크하는 함수
